Struct a2::client::Client

pub struct Client { /* fields omitted */ }

Handles requests to and responses from Apple Push Notification service. Connects using a given connector. Handles the needed authentication and maps responses.

The send method returns a future, which is successful when APNs receives the notification and responds with a status OK. In any other case the future fails. If APNs gives a reason for the failure, the returned Err holds the response for handling.

Methods

impl Client

pub fn certificate<R>(
certificate: &mut R,
password: &str,
endpoint: Endpoint
) -> Result<Client, Error> where
R: Read,

Create a connection to APNs using the provider client certificate which you obtain from your Apple developer account.

pub fn token<S, T, R>(
pkcs8_pem: R,
key_id: S,
team_id: T,
endpoint: Endpoint
) -> Result<Client, Error> where
S: Into<String>,
T: Into<String>,
R: Read,

Create a connection to APNs using system certificates, signing every request with a signature using a private key, key id and team id provisioned from your Apple developer account.

pub fn send<'a>(&self, payload: Payload<'a>) -> FutureResponse

Send a notification payload. Returns a future that needs to be given to an executor.

pub fn send_with_timeout<'a>(
&self,
message: Payload<'a>,
timeout: Duration
) -> Timeout<FutureResponse>

Sends a notification with a timeout. Triggers Error::TimeoutError if the request takes too long.