Module a2::request::notification

The aps notification content builders

Structs

CollapseId

LocalizedAlert

LocalizedNotificationBuilder
A builder to create a localized APNs payload.

NotificationOptions
Headers to specify options to the notification.

PlainNotificationBuilder
A builder to create a simple APNs notification payload.

SilentNotificationBuilder
A builder to create an APNs silent notification payload which can be used to send custom data to the user's phone if the user hasn't been running the app for a while. The custom data needs to be implementing Serialize from Serde.

Enums

Priority
The importance how fast to bring the notification for the user..

Traits

NotificationBuilder